--- src/app/dashboard/employees/new/page.tsx
@@ -1,18 +1,0 @@
-import { Box, Heading } from '@/ui';
-
-import { BackButton } from './BackButton';
-import { NewEmployeeForm } from './NewEmployeeForm';
-
-export default async function Page() {
-  return (
-    <Box>
-      <Box className="py-5 flex flex-col gap-[40px]">
-        <Heading as="h1" variant="text-xl">New Employee</Heading>
-      </Box>
-      <Box className="py-5 flex flex-col gap-[40px]">
-        <BackButton />
-        <NewEmployeeForm />
-      </Box>
-    </Box>
-  );
-}
