+++ src/pages-components/Dashboard/EditEmployeePage/index.ts
@@ -1,0 +1,1 @@
+export { EditEmployeePage } from './EditEmployeePage';
