+++ src/pages-components/Dashboard/EmployeesPage/index.ts
@@ -1,0 +1,1 @@
+export { EmployeesPage } from './EmployeesPage';
