+++ src/pages-components/Dashboard/NewEmployeePage/index.ts
@@ -1,0 +1,1 @@
+export { NewEmployeePage } from './NewEmployeePage';
