+++ src/pages-components/Dashboard/NewEmployeePage/schema.ts
