+++ src/pages-components/Dashboard/SingleEmployeePage/index.ts
@@ -1,0 +1,1 @@
+export { SingleEmployeePage } from './SingleEmployeePage';
