+++ src/pages/dashboard/employees/[id]/edit.ts
@@ -1,0 +1,1 @@
+export { EditEmployeePage as default } from '@/pages-components/Dashboard/EditEmployeePage';
