+++ src/pages/dashboard/employees/[id]/index.ts
@@ -1,0 +1,1 @@
+export { SingleEmployeePage as default } from '@/pages-components/Dashboard/SingleEmployeePage';
