+++ src/pages/dashboard/employees/index.ts
@@ -1,0 +1,1 @@
+export { EmployeesPage as default } from '@/pages-components/Dashboard/EmployeesPage';
