+++ src/pages/dashboard/employees/new.ts
@@ -1,0 +1,1 @@
+export { NewEmployeePage as default } from '@/pages-components/Dashboard/NewEmployeePage';
